Add 32-bit support
Converting TimeInterval object to milliseconds exceeds Int32.max,
which would crash 32-bit devices such as iPhone 5. Apparently,
crashes related to hash value conversion from String is also
caused by the same issue.

The fix is to use Int64 where the values would exceed Int32.max,
mainly in timestamp calculation and hash calculation.

+++ Source/Logging/ExposureLog.swift
@@ -17,7 +17,7 @@
 /// https://facebook.github.io/planout/docs/logging.html
 public struct ExposureLog: Encodable {
     public let event: String = "exposure"
-    public let timestampInMs: Int
+    public let timestampInMs: Int64
     public let name: String
     public let salt: String
     public let checksum: String
@@ -40,7 +40,7 @@ public struct ExposureLog: Encodable {
         checksum = experiment.definition.checksum
 
         // Swift TimeInterval are in seconds. Multiply by 1000 to convert to milliseconds.
-        timestampInMs = Int(Date().timeIntervalSince1970 * 1000)
+        timestampInMs = Int64(Date().timeIntervalSince1970 * 1000)
     }
 
     private enum CodingKeys: String, CodingKey {
